Disable pprof routes in production build
