+++ server/config.go
@@ -47,7 +47,6 @@ type Config struct {
 	SourceMapRaw        json.RawMessage        `json:"sourceMap"`
 	CompressRaw         json.RawMessage        `json:"compress"`
 	LegacyServer        string                 `json:"legacyServer"` // normally you don't need to set this
-	Pprof               bool                   `json:"pprof"`
 	Minify              bool                   `json:"-"`
 	SourceMap           bool                   `json:"-"`
 	Compress            bool                   `json:"-"`
