+++ server/release.go
@@ -4,6 +4,8 @@ package server
 
 import (
 	"embed"
+
+	"github.com/ije/rex"
 )
 
 // production mode
@@ -14,3 +16,10 @@ var embedFS embed.FS
 
 // may be changed by `-ldflags`
 var VERSION = "v136"
+
+// pprof is disabled in production build
+func pprofRouter() rex.Handle {
+	return func(ctx *rex.Context) any {
+		return rex.Next()
+	}
+}
